refactor(recovery): Use base32 for recovery code generation
