+++ lib/db/util.js
@@ -4,22 +4,11 @@
 
 'use strict'
 
-const assert = require('assert')
 const crypto = require('crypto')
 const P = require('../promise')
 const randomBytes = P.promisify(require('crypto').randomBytes)
 const scryptHash = P.promisify(require('scrypt-hash'))
-const random = require('./random')
-
-const ALPHABET = '0123456789ABCDEFGHJKMNPQRSTVWXYZ'
-
-// some sanity checks, hard to test, private to this mdoule
-assert.equal(ALPHABET.length, 32, 'ALPHABET is 32 characters')
-assert.equal(ALPHABET.indexOf('I'), -1, 'should not contain I')
-assert.equal(ALPHABET.indexOf('L'), -1, 'should not contain L')
-assert.equal(ALPHABET.indexOf('O'), -1, 'should not contain O')
-assert.equal(ALPHABET.indexOf('U'), -1, 'should not contain U')
-
+const base32 = require('./random')
 
 const BOUNCE_TYPES = new Map([
   ['__fxa__unmapped', 0], // a bounce type we don't yet recognize
@@ -53,18 +42,6 @@ const VERIFICATION_METHODS = new Map([
   ['recovery-code', 3]   // Recovery code
 ])
 
-function base32(len) {
-  return random(len)
-    .then(bytes => {
-      const out = []
-
-      for (let i = 0; i < len; i++) {
-        out.push(ALPHABET[bytes[i] % 32])
-      }
-
-      return out.join('')
-    })
-}
 
 module.exports = {
 
@@ -130,9 +107,7 @@ module.exports = {
     }
 
     return P.all(randomCodes)
-      .then((result) => {
-        return result
-      })
+
   },
 
   // A helper function for aggregating name:value pairs into a JSON object.
